requirements.txt
amqp==5.3.1
argcomplete==3.6.3
asgiref==3.10.0
billiard==4.2.3
bleach==6.3.0
boto3==1.41.5
botocore==1.41.5
celery==5.3.1
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
commitizen==4.10.0
decli==0.6.3
Deprecated==1.3.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-ranged-response==0.2.0
django-redis==6.0.0
django-simple-captcha==0.6.2
django-storages==1.14.6
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
flower==2.0.1
gunicorn==23.0.0
humanize==4.14.0
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.6.1
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
questionary==2.1.1
redis==7.1.0
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
termcolor==3.2.0
tomlkit==0.13.3
tornado==6.5.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
wrapt==2.0.1
